requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
geoalchemy2==0.6.3
joblib==0.14.0            # via scikit-learn
numpy==1.17.3             # via pandas, scikit-learn, scipy
pandas==0.25.2
plotly==4.2.1
psycopg2-binary==2.8.4
python-dateutil==2.8.0    # via pandas
pytz==2019.3              # via pandas
pyyaml==5.1.2
retrying==1.3.3           # via plotly
scikit-learn==0.21.3
scipy==1.3.1              # via scikit-learn
simplejson==3.16.0
six==1.12.0               # via plotly, python-dateutil, retrying, sqlalchemy-utils
sqlalchemy-utils==0.34.2
sqlalchemy==1.3.10
trafaret==2.0.0
